FT: httpd: upgrade.c: fix compiler warnings
Update from FreshTomato
https://github.com/pedro0311/freshtomato-arm/commit/dcabe7fed8f173492bafddb968c706d451d097be

+++ tomato64/package/libshared/shared/shutils.c
@@ -492,10 +492,10 @@ safe_fread(void *ptr, size_t size, size_t nmemb, FILE *stream)
  * @param	stream	file stream
  * @return	number of items successfully written
  */
-size_t
-safe_fwrite(const void *ptr, size_t size, size_t nmemb, FILE *stream)
+int
+safe_fwrite(const void *ptr, size_t size, int nmemb, FILE *stream)
 {
-	size_t ret = 0;
+	int ret = 0;
 
 	do {
 		clearerr(stream);
